+++ l3build-check.lua
@@ -798,7 +798,8 @@ function runtest(name, engine, hide, ext, test_type, breakout)
     format = " --fmt=" .. format
   end
   -- Special casing for XeTeX engine
-  if match(engine, "xetex") and test_type.generated ~= pdfext then
+  if match(engine, "xetex") and test_type.generated ~= pdfext and
+      xetexnopdf then
     checkopts = checkopts .. " -no-pdf"
   end
   -- Special casing for ConTeXt